+++ func_profiler/func_profiler_simple_minmax.go
@@ -42,24 +42,24 @@ func NewSimpleMinMaxProfiler() *simpleMinMaxProfiler {
 }
 
 // Register a function for profiling
-func (p *simpleMinMaxProfiler) RegisterFunc(name string) uint16 {
+func (p *simpleMinMaxProfiler) RegisterFunc(name string) int {
 	id := p.base.RegisterFunc(name)
 	p.accum = append(p.accum, newSimpleMinMaxAccum())
 	return id
 }
 
 // Clear the data associated with a function
-func (p *simpleMinMaxProfiler) ClearData(id uint16) {
+func (p *simpleMinMaxProfiler) ClearData(id int) {
 	p.accum[id] = newSimpleMinMaxAccum()
 }
 
 // Start timing a function for simpleMinMaxProfiler
-func (p *simpleMinMaxProfiler) StartTimer(id uint16) {
+func (p *simpleMinMaxProfiler) StartTimer(id int) {
 	p.base.StartTimer(id)
 }
 
 // Stop timing a function for simpleMinMaxProfiler
-func (p *simpleMinMaxProfiler) EndTimer(id uint16) {
+func (p *simpleMinMaxProfiler) EndTimer(id int) {
 	ms := p.base.EndTimer(id)
 	accum := &p.accum[id]
 	accum.totalTime += ms
@@ -73,23 +73,23 @@ func (p *simpleMinMaxProfiler) EndTimer(id uint16) {
 }
 
 // Get the average runtime for a function
-func (p *simpleMinMaxProfiler) GetAvg(id uint16) float64 {
+func (p *simpleMinMaxProfiler) GetAvg(id int) float64 {
 	return (p.accum[id].totalTime /
 		float64(p.accum[id].invocations))
 }
 
 // Get the name of a given function
-func (p *simpleMinMaxProfiler) GetName(id uint16) string {
+func (p *simpleMinMaxProfiler) GetName(id int) string {
 	return p.base.GetName(id)
 }
 
 // Set the name of a given function
-func (p *simpleMinMaxProfiler) SetName(id uint16, name string) {
+func (p *simpleMinMaxProfiler) SetName(id int, name string) {
 	p.base.SetName(id, name)
 }
 
 // Return a string displaying the stats for a given function
-func (p *simpleMinMaxProfiler) GetSummaryString(id uint16) string {
+func (p *simpleMinMaxProfiler) GetSummaryString(id int) string {
 	return fmt.Sprintf("Summary for %s: {Avg: %.3f}",
 		p.base.GetName(id),
 		p.GetAvg(id))
